improvements to moves
Normalization copies original event properties. This prevents the
browser's event side-effects from effecting the touch "higher-order"
events.

"movesAsObservable" will now include the computed "down" event as part
of the moves group. This allows delta-scanning to work properly, since
without it the first "delta" is from the first move to the second move,
instead of the origin to the first move. Complicated. If you need
explaination just ask. Otherwise, just trust this is for the best. :)